ConDec-446: Move CommentSplitter class to extraction package
